Put all enums into one class (to link them to the subject)
