+++ src/main/java/net/sf/mzmine/modules/peaklistmethods/io/spectraldbsubmit/MSPEntryGenerator.java
@@ -31,7 +31,7 @@ package net.sf.mzmine.modules.peaklistmethods.io.spectraldbsubmit;
 
 import net.sf.mzmine.datamodel.DataPoint;
 import net.sf.mzmine.modules.peaklistmethods.identification.spectraldbsearch.dbentry.DBEntryField;
-import net.sf.mzmine.modules.peaklistmethods.io.spectraldbsubmit.LibrarySubmitParameters.Polarity;
+import net.sf.mzmine.modules.peaklistmethods.io.spectraldbsubmit.GnpsValues.Polarity;
 
 public class MSPEntryGenerator {
